requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pymysql
python-dotenv
insightface
onnxruntime
opencv-python
numpy
Pillow
